change font, refactor my info
