--- src/Style.scss
@@ -1,15 +1,0 @@
-@mixin project{
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: start;
-  margin-bottom: 40px;
-  padding: 5px 10px 5px 10px;
-
-  &:hover {
-    border-radius: 10px;
-    background-color: rgba(128, 128, 128, 0.3);
-    cursor: pointer;
-  }
-}
-
