--- src/components/Experience.scss
@@ -1,6 +1,0 @@
-@use "../Style.scss" as style;
-
-
-.individual-experience {
-  @include style.project;
-}
