--- src/components/Project.scss
@@ -1,6 +1,0 @@
-@use "../Style.scss" as style;
-
-
-.individual-project {
-  @include style.project;
-}
