+++ src/components/ProjectDescription.scss
@@ -1,6 +1,8 @@
 .skill {
+  color: #fff066;
+  margin-top: 7px;
   margin-right: 7px;
-  background-color: gray;
+  background-color: rgba(217,217,217, 0.2);
   padding: 2px 5px 2px 5px;
   border-radius: 20px;
 }
