+++ src/components/home/Home.scss
@@ -8,10 +8,10 @@ $colorForName: rgb(232, 211, 162);
   font-size: 2rem;
 }
 
-p {
+.home p {
   margin: 0;
   padding: 0;
-
+  color: white;
 }
 
 #full-name {
