--- src/data/experience.js
@@ -1,44 +1,0 @@
-const experiences = [
-  {
-    title: "Software Engineer Intern",
-    company: "Xemelgo",
-    date: {
-      start: "October 2019",
-      end: "March 2020"
-    },
-    descriptions: [
-      "Administered the end-to-end functions of a new job tracking feature to aid customers in tracing actual labor hours.",
-      "Successfully designed and deployed DynamoDB tables and RESTful APIs on AWS through utilization of Serverless framework.",
-      "Leveraged React Native to support mobile development and simulated on Android and iOS platforms."
-    ],
-    skills: ["JavaScript", "React", "HTML/CSS", "AWS"]
-  },
-  {
-    title: "Undergraduate Research Assistant",
-    company: "University of Washington",
-    date: {
-      start: "January 2020",
-      end: "December 2019"
-    },
-    descriptions: [
-      "Improved the quality of a national curb dataset by assessing a sample of data points, identifying locations with incorrect curb data, and creating a data validation tool to improve the data accuracy in the future.",
-      "Established backend utilizing NodeJS, ExpressJS, Turf to process GeoJSON files from public city data."
-    ],
-    skills: ["JavaScript", "React", "NodeJS"]
-  },
-  {
-    title: "Undergraduate Research Assistant",
-    company: "South Seattle College",
-    date: {
-      start: "January 2017",
-      end: "March 2017"
-    },
-    descriptions: [
-      "Deployed a Raspberry Pi sensor to outline wifi traffic and to insert data in MongoDB database for analysis.",
-      "Utilized software package or analysis toolset to turn raw data into more useful information and insights."
-    ],
-    skills: ["Python", "MongoDB", "Jupiter"]
-  }
-]
-
-export default experiences;
