--- src/data/project.js
@@ -1,42 +1,0 @@
-const projects = [
-  {
-    title: "Portfolio Website",
-    date: {
-      start: "December 2023",
-      end: "Present"
-    },
-    descriptions: [
-      "Develop personal responsive website to showcase my portfolio of software development projects.",
-      "Optimize website performance and usability with SEO best practices in mind.",
-    ],
-    skills: ["JavaScript", "React", "HTML/CSS", "SCSS", "Bootstrap", "MUI"]
-  },
-  {
-    title: "Collaborative Canvas",
-    date: {
-      start: "December 2023",
-      end: "January 2024"
-    },
-    descriptions: [
-      "Utilized technical skills in software development to innovate an interactive tool for seamless collaboration and demonstration within the team.",
-      "Empower teams to improved productivity and visualized creation in remote environment.",
-      "Designed and developed a robust and scalable web application by implementing industry best practices."
-    ],
-    skills: ["JavaScript", "React", "HTML/CSS", "SCSS", "Bootstrap", "NodeJS", "Socket.IO"]
-  },
-  {
-    title: "Stock Performance",
-    date: {
-      start: "January 2018",
-      end: "March 2018"
-    },
-    descriptions: [
-      "Permitted stock investors to compare various stocks’ performance.",
-      "Extracted stock data from Quandl financial services API and inserted into MySQL database.",
-      "Adopted Object-oriented design and MVC, organizing the code for maintainability and scalability."
-    ],
-    skills: ["Python", "MongoDB", "Jupiter"]
-  }
-]
-
-export default projects;
