+++ src/index.scss
@@ -5,14 +5,16 @@
 }
 
 body {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+  font-family: 'Inter';font-size: 22px;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background-color: rgb(51, 0, 111);
-  color: rgb(255, 255, 255);
+  color: rgb(217,217,217);
   font-size: 15px;
+}
+
+h1, h2, h3, h4, h5 {
+  color: white;
 }
 
 code {
